Disable the EnableTargetOSChecking language option for all Swift AST contexts.
LLDB started to understand the minimum deployment target embedded in
the Mach-O load commands in r332067. This makes it more likely to hit
the module-too-new warning in a multi-dylib project where any of the
dylibs have a newer minimum deployment target than the main
executable. Since this is only a warning emitted by the serialized
module loader, there isn't much to be gained to enable this warning as
error in LLDB. This patch disables it.

rdar://problem/41652444

apple-llvm-split-commit: a72289ceb4f23d6c7e7b667b70cb84eac4dffa61
apple-llvm-split-dir: lldb/