[IRInterpreter] Fix misevaluation of interpretation expressions with `urem`.
Scalar::MakeUnsigned was implemented incorrectly so it didn't
really change the sign of the type (leaving signed types signed).
This showed up as a misevaluation when IR-interpreting urem but
it's likely to arise in other contexts.

This commit fixes the definition, and adds a test to make
sure this won't regress in future (hopefully).

Fixes rdar://problem/42038760 and LLVM PR38076

Differential Revision:  https://reviews.llvm.org/D49155

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@336872 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: 963ca1ff7f96380d340bfd18443c8464a3edb4fe
apple-llvm-split-dir: lldb/