Fix the libcxx set, multiset, vector and bitset formatters to work on references.
The synthetic child providers for these classes had a type expression that matched
pointers & references to the type, but the Front End only worked on the actual object.

I fixed this by adding a way for the Synthetic Child FrontEnd provider to request dereference,
and then had these formatters use that mode.

<rdar://problem/40849836>

Differential Revision: https://reviews.llvm.org/D49279


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@337035 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: ef7307f42ca3fa19c9ea559d355ebef0350b4891
apple-llvm-split-dir: lldb/