Disable EnableTargetOSChecking in the scratch AST context.
This is not needed in the general case, as this flag is set by the
expression parser. There is one case where it matters though: if the
Swift code is loaded by a main executable that has a deploymnent target
lower than 10.9, then we will fail to import the standard library
module in SwiftASTContext::CreateInstance().

I'm still working on a testcase for this, but I wanted to get the trivial
fix in for users to test.

(cherry picked from commit 49922e9da77631db6e60773e7674cbe7535f959b)

apple-llvm-split-commit: 3b219579f53d4d7557656c7781ff1345144e9f62
apple-llvm-split-dir: lldb/