Delay passing the persistent_state till after making the expression could invalidate it.
In the SwiftExpressionParser, we set the persistent_state associated with the currently
active scratch AST context into a couple of delegate objects that we will use at the
end of the expression evaluation.  But in the course of making an expression parser in the context
of a new module, some conflicting imports could invalidate the scratch AST context, leaving
the persistent_state pointers as dangling pointers.

So this patch delays getting and passing the persistent_state to the delegates till after
making the new SwiftExpressionParser, at which point we know the ones we calculate are the
ones we really intend to use.

Before this patch, I was getting a crash in:

lang/swift/clangimporter/objcmain_conflicting_dylibs_bridging_headers

<rdar://problem/42410482>

apple-llvm-split-commit: 4254d862f02b9397003a09f4bae1d6dbc7822c77
apple-llvm-split-dir: lldb/