[cmake] LLDBFramework: Copy in the clang resource dir from the swift build
This module attempted to copy in a clang resource dir into the LLDB
framework from a non-existent directory (at least, this is the case when
building & testing with build-script --lldb-build-with-cmake --test).

Copy in an actual clang resource directory which is known to live within
the swift build.

This makes it possible to run tests that do anything interesting with
the clang resource dir (e.g importing C modules).

apple-llvm-split-commit: c333ba5034a76bf6b1815851e12176cd7a91455f
apple-llvm-split-dir: lldb/