[X86] In EFLAGS copy pass, don't emit EXTRACT_SUBREG instructions since we're after peephole
Normally the peephole pass converts EXTRACT_SUBREG to COPY instructions. But we're after peephole so we can't rely on it to clean these up.

To fix this, the eflags pass now emits a COPY with a subreg input.

I also noticed that in 32-bit mode we need to constrain the input to the copy to ensure the subreg is valid. Otherwise we'll fail verify-machineinstrs

Differential Revision: https://reviews.llvm.org/D50656

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@339945 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: ee2d39c54b00bf268d0697d9a6ee988512ff8ee7
apple-llvm-split-dir: llvm/