Support swifterror in coroutine lowering.
The support for swifterror allocas should work in all lowerings.
The support for swifterror arguments only really works in a lowering
with prototypes where you can ensure that the prototype also has a
swifterror argument; I'm not really sure how it could possibly be
made to work in the switch lowering.

apple-llvm-split-commit: 83980d7d6933feffced1b9bcbde63cf32ebcb3f9
apple-llvm-split-dir: llvm/