[analyzer] Drop support for GC mode in RetainCountChecker
A lot of code in RetainCountChecker deals with GC mode.
Given that GC mode is deprecated, Apple does not ship runtime for it,
and modern compiler toolchain does not support it, it makes sense to
remove the code dealing with it in order to aid understanding of
RetainCountChecker.

Differential Revision: https://reviews.llvm.org/D50747

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@340091 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: f9cb5d7bb5c4fa645d79b8dfeafebbfa2888a369
apple-llvm-split-dir: clang/