[DebugInfo] Have the verifier accept missing linkage names.
According to the standard, for the .debug_names (the "dwarf accelerator
tables"):

> If a subprogram or inlined subroutine is included, and has a
> DW_AT_linkage_name attribute, there will be an additional index entry
> for the linkage name.

For Swift we generate DW_structure_types with a linkage name and the
verifier was incorrectly rejecting this. This patch fixes that by only
considering the linkage name in those particular cases. The test is the
"reduced" debug info of the failing swift test on swift.org.

Differential revision: https://reviews.llvm.org/D51420

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@341311 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit cccbfeb670418db662435dd95c58869325fa046b)

apple-llvm-split-commit: 64d62f60fbadee4fe8373666e7c129646beb4195
apple-llvm-split-dir: llvm/