NFC: more memset inline arm64 coverage
I'm looking at some codegen optimization in this area and want to make sure I understand the current codegen and don't regress it. This patch further expands the tests (which I already expanded in r341406) to capture more of the current code generation when it comes to stack-based small non-zero memset on arm64. This patch annotates some potential fixes.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@341493 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: f6640ab701521a7cda832bbe2291cfc693e434d2
apple-llvm-split-dir: llvm/