IndexStore: repair Windows build
Building IndexStore would fail as `dllimport` would be applied to a non-inlined
function definition which is not permitted.  Remove the declaration to change
the error to a warning (inconsistent dll storage).  This allows us to build
correctly.  We could always apply the `dllexport` attribute to the definitions
to avoid the warning.

apple-llvm-split-commit: 49423a26114a213b2e1096a4c0a02d881311598a
apple-llvm-split-dir: clang/