IndexStore: correct symbol visibility
On ELF targets, use protected visibility (symbols cannot be interpositioned).
On MachO, use default visibility.  On Windows, we need to define it to either
`__declspec(dllexport)` when building the library itself, or
`__declspec(dllimport)` when using the library.  This fixes a number of warnings
when building IndexCore for Windows.

Remove the use of the exports list and rely on the annotations instead.  This
makes the build similar across all the targets (Windows requires all listed
symbols to be present, and we would need to preprocess the list otherwise).

Fix the accidental usage of `_CINDEX_LIB_` opting for the CMake standard
variable `<Name>_EXPORTS`.

Simplify the logic in the CMake for Apple targets a bit.

apple-llvm-split-commit: 4a0fee25e743405748439e579eec577fb7ad9441
apple-llvm-split-dir: clang/