[analyzer][NFC] Collect all -analyzer-config options in a .def file
I'm in the process of refactoring AnalyzerOptions. The main motivation behind
here is to emit warnings if an invalid -analyzer-config option is given from the
command line, and be able to list them all.

In this patch, I'm moving all analyzer options to a def file, and move 2 enums
to global namespace.

Differential Revision: https://reviews.llvm.org/D53277

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@345986 91177308-0d34-0410-b5e6-96231b3b80d8
apple-llvm-split-commit: f7e21ce46e5a47c84f0835bade6335702041aead
apple-llvm-split-dir: clang/