Fix a use-after-free introduced by r344915.
r344915 added a call to ApplyDebugLocation to the sanitizer check
function emitter. Some of the sanitizers are emitted in the function
epilogue though and the LexicalScopeStack is emptied out before. By
detecting this situation and early-exiting from ApplyDebugLocation the
fallback location is used, which is equivalent to the return location.

rdar://problem/45859802

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@346454 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 130d3d039ee8886084c8dfe038b3183db72f6788)

apple-llvm-split-commit: 35c7180c4715f15f40c335626c45586dd00f9ad0
apple-llvm-split-dir: clang/